[CWS] extract `evalRule` and `checkPoliciesLocal` to common package
