+++ cmd/system-probe/subcommands/runtime/activity_dump.go
@@ -627,3 +627,10 @@ func stopActivityDump(_ log.Component, _ config.Component, _ secrets.Component, 
 	fmt.Println("done!")
 	return nil
 }
+
+func printStorageRequestMessage(prefix string, storage *api.StorageRequestMessage) {
+	fmt.Printf("%so file: %s\n", prefix, storage.GetFile())
+	fmt.Printf("%s  format: %s\n", prefix, storage.GetFormat())
+	fmt.Printf("%s  storage type: %s\n", prefix, storage.GetType())
+	fmt.Printf("%s  compression: %v\n", prefix, storage.GetCompression())
+}
